+++ backend/src/baserow/contrib/database/table/handler.py
@@ -257,7 +257,9 @@ class TableHandler:
 
         field_options = {notes.id: {"width": 400}, active.id: {"width": 100}}
         fields = [notes, active]
-        view_handler.update_field_options(user, view, field_options, fields=fields)
+        view_handler.update_field_options(
+            user=user, view=view, field_options=field_options, fields=fields
+        )
 
         model = table.get_model(attribute_names=True)
         model.objects.create(name="Tesla", active=True, order=1)